V0.3

+++ domain/src/mongo.js
@@ -2,19 +2,24 @@ const mongodb = require('mongodb');
 const mongo = mongodb.MongoClient;
 const ObjectID = mongodb.ObjectID;
 
-const url = 'mongodb://localhost:27017/kladovka';
-const coll = 'items';
+
+async function connect(url) {
+    return mongo.connect(url);
+}
+
+function disconnect(db) {
+    db.close();
+}
 
 /**
  * Сохраняет объект в БД
  * @param {Object} item - Этот объект будет сохранен в базу
  * @returns {Promise.<String, Error>} id добавленного объекта
  */
-async function addItem(item) {
-    let db = await mongo.connect(url);
+async function addItem(db, coll, item) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.insertOne(item);
-    db.close();
     return res.insertedId;
 }
 
@@ -23,11 +28,10 @@ async function addItem(item) {
  * @param {String} id - идентификатор искомого объекта
  * @returns {Promise.<Object, Error>} Объект или null если такого объекта нет
  */
-async function getNotDeletedItemById(id) {
-    let db = await mongo.connect(url);
+async function getNotDeletedItemById(db, coll, id) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let item = await collection.findOne({ _id: ObjectID(id), deleted: undefined });
-    db.close();
     return item;
 }
 
@@ -36,11 +40,10 @@ async function getNotDeletedItemById(id) {
  * @param {String} id - идентификатор удаляемого объекта
  * @returns {Promise.<String, Error>} id удалённого объекта
  */
-async function deleteItemById(id) {
-    let db = await mongo.connect(url);
+async function deleteItemById(db, coll, id) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.updateOne({ _id: ObjectID(id) }, { $set: { 'deleted': true } });
-    db.close();
     if (res.result.nModified === 0) return null;
     return id;
 }
@@ -49,11 +52,10 @@ async function deleteItemById(id) {
  * Удаляет ВСЕ объекты из БД
  * @returns {Promise.<Number, Error>} Количество удалённых объектов
  */
-async function deleteAllItems() {
-    let db = await mongo.connect(url);
+async function deleteAllItems(db, coll) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.updateMany({ deleted: undefined }, { $set: { deleted: true } });
-    db.close();
     return res.modifiedCount;
 }
 
@@ -61,11 +63,10 @@ async function deleteAllItems() {
  * Получает массив объектов(не удалённых)
  * @returns {Promise.<Array, Error>} Массив объектов
  */
-async function getNotDeletedItems() {
-    let db = await mongo.connect(url);
+async function getNotDeletedItems(db, coll) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.find({ 'deleted': undefined }).toArray();
-    db.close();
     return res;
 }
 
@@ -73,11 +74,10 @@ async function getNotDeletedItems() {
  * Получает массив объектов(не удалённых) данного типа
  * @returns {Promise.<Array, Error>} Массив объектов
  */
-async function getAllItemsByType(type) {
-    let db = await mongo.connect(url);
+async function getAllItemsByType(db, coll, type) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.find({ deleted: undefined, type }).toArray();
-    db.close();
     return res;
 }
 
@@ -85,15 +85,16 @@ async function getAllItemsByType(type) {
  * Очищает коллекцию
  * @returns {Promise.<Number, Error>} Количество удалённых объектов
  */
-async function clearCollection() {
-    let db = await mongo.connect(url);
+async function clearCollection(db, coll) {
+    if (db === undefined) throw new Error('нет базы данных');
     let collection = db.collection(coll);
     let res = await collection.deleteMany({});
-    db.close();
     return res.deletedCount;
 }
 
 module.exports = {
+    connect,
+    disconnect,
     add: addItem,
     deleteById: deleteItemById,
     deleteAll: deleteAllItems,
